+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## [next]
+
+Fixes:
+
+  - `XapianDb.database.delete_docs_of_class(klass)` now deletes docs of all descendants of `klass`, too (if tracked)
+
 ##1.3.6 (August 14th, 2015)
 
 Changes: